+++ .gitignore
@@ -1,12 +1,19 @@
+# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+
 # dependencies
-node_modules
-.pnp
+/node_modules
+/.pnp
 .pnp.js
 
+# testing
+/coverage
+
 # next.js
-website/.next/
-website/out/
-website/build
+/.next/
+/out/
+
+# production
+/build
 
 # misc
 .DS_Store
