+++ .vscode/settings.json
@@ -2,6 +2,5 @@
   "editor.formatOnSave": true,
   "editor.codeActionsOnSave": {
     "source.fixAll.eslint": true
-  },
-  "eslint.workingDirectories": ["./website"]
+  }
 }
