+++ src/app/page.tsx
@@ -1,0 +1,7 @@
+export default function HomePage() {
+  return (
+    <main>
+      <h1>Home Page</h1>
+    </main>
+  );
+}
