--- website/src/app/page.tsx
@@ -1,9 +1,0 @@
-import { Heading } from "~/libs/chakra-ui/react";
-
-export default function HomePage() {
-  return (
-    <main>
-      <Heading>Home Page</Heading>
-    </main>
-  );
-}
