--- website/src/app/providers.tsx
@@ -1,9 +1,0 @@
-import ChakraUIProvider from "~/libs/chakra-ui/config/ThemeProvider";
-
-interface ProvidersProps {
-  children: React.ReactNode;
-}
-
-export default function Providers({ children }: ProvidersProps) {
-  return <ChakraUIProvider>{children}</ChakraUIProvider>;
-}
