--- website/src/libs/chakra-ui/config/ThemeProvider.tsx
@@ -1,14 +1,0 @@
-import { CacheProvider } from "../next-js";
-import { ChakraProvider } from "../react";
-
-interface Props {
-  children: React.ReactNode;
-}
-
-export default function ChakraUIProvider({ children }: Props) {
-  return (
-    <CacheProvider>
-      <ChakraProvider>{children}</ChakraProvider>
-    </CacheProvider>
-  );
-}
